Refactor document search into a normal package and an extension.
