+++ packages/csvviewer-extension/package.json
@@ -34,6 +34,7 @@
     "@jupyterlab/apputils": "^0.19.1",
     "@jupyterlab/csvviewer": "^0.19.1",
     "@jupyterlab/docregistry": "^0.19.1",
+    "@jupyterlab/documentsearch": "^0.7.1",
     "@jupyterlab/documentsearch-extension": "^0.19.1",
     "@jupyterlab/mainmenu": "^0.8.1",
     "@phosphor/datagrid": "^0.1.6",
