+++ packages/csvviewer-extension/tsconfig.json
@@ -19,6 +19,9 @@
       "path": "../docregistry"
     },
     {
+      "path": "../documentsearch"
+    },
+    {
       "path": "../documentsearch-extension"
     },
     {
