+++ packages/documentsearch-extension/package.json
@@ -34,6 +34,7 @@
     "@jupyterlab/codeeditor": "^0.19.1",
     "@jupyterlab/codemirror": "^0.19.1",
     "@jupyterlab/docregistry": "^0.19.1",
+    "@jupyterlab/documentsearch": "^0.7.1",
     "@jupyterlab/mainmenu": "^0.8.1",
     "@jupyterlab/notebook": "^0.19.2",
     "@phosphor/coreutils": "^1.3.0",
