--- packages/documentsearch-extension/src/searchproviders.ts
@@ -1,4 +1,0 @@
-// Copyright (c) Jupyter Development Team.
-// Distributed under the terms of the Modified BSD License.
-export * from './providers/notebooksearchprovider';
-export * from './providers/codemirrorsearchprovider';
