+++ packages/documentsearch-extension/tsconfig.json
@@ -25,6 +25,9 @@
       "path": "../docregistry"
     },
     {
+      "path": "../documentsearch"
+    },
+    {
       "path": "../mainmenu"
     },
     {
