+++ packages/documentsearch/package.json
@@ -27,6 +27,19 @@
     "prepublishOnly": "npm run build",
     "watch": "tsc -w --listEmittedFiles"
   },
+  "dependencies": {
+    "@jupyterlab/apputils": "^0.19.1",
+    "@jupyterlab/cells": "^0.19.1",
+    "@jupyterlab/codeeditor": "^0.19.1",
+    "@jupyterlab/codemirror": "^0.19.1",
+    "@jupyterlab/notebook": "^0.19.2",
+    "@phosphor/coreutils": "^1.3.0",
+    "@phosphor/disposable": "^1.1.2",
+    "@phosphor/signaling": "^1.2.2",
+    "@phosphor/widgets": "^1.6.0",
+    "codemirror": "~5.42.0",
+    "react": "~16.4.2"
+  },
   "devDependencies": {
     "rimraf": "~2.6.2",
     "typescript": "~3.3.1"
