+++ packages/documentsearch/src/index.ts
@@ -1,4 +1,6 @@
-/*-----------------------------------------------------------------------------
-| Copyright (c) Jupyter Development Team.
-| Distributed under the terms of the Modified BSD License.
-|----------------------------------------------------------------------------*/
+// Copyright (c) Jupyter Development Team.
+// Distributed under the terms of the Modified BSD License.
+
+export * from './searchproviderregistry';
+export * from './searchinstance';
+export * from './interfaces';
