+++ packages/documentsearch/src/providers/codemirrorsearchprovider.ts
