+++ packages/documentsearch/src/providers/notebooksearchprovider.ts
