+++ packages/documentsearch/src/searchoverlay.tsx
