+++ packages/documentsearch/style/index.css
